requirements.txt
Flask
google-generativeai
weasyprint
python-dotenv
google-adk
